rename GetJobRequirementText to GetJobRequirements
